Implement a simple scorer
So that we can generate harder puzzle for human.

TODO: implement more strategy.

+++ build.sh
@@ -7,8 +7,10 @@ cd wasm
 wasm-pack build --target web
 cd ..
 npm run compile
-cd build
-ln -s ../wasm .
-cd ..
+if [ ! -e build/wasm ]; then
+  cd build
+  ln -s ../wasm .
+  cd ..
+fi
 npx sass scss/:build/
 python3 -m http.server 8080
